# Use objects instead of classes to represent "defined" types