--- a/lib/graphql/client/definition.rb
+++ b/lib/graphql/client/definition.rb
@@ -144,8 +144,8 @@ module GraphQL
 
         def cast_object(obj)
           if obj.class.is_a?(GraphQL::Client::Schema::ObjectType)
-            unless obj.class._spreads.include?(definition_node.name)
-              raise TypeError, "#{definition_node.name} is not included in #{obj.class.source_definition.name}"
+            unless obj._spreads.include?(definition_node.name)
+              raise TypeError, "#{definition_node.name} is not included in #{obj.source_definition.name}"
             end
             schema_class.cast(obj.to_h, obj.errors)
           else