… test/test_query_result.rb