Replacing #213

This attempts to improve the memory usage of object type definitions in applications with a large number of queries. This is done by having `define_class` return a plain object instead of a dynamic subclass.

The behaviour is a bit confusing here because we are building a GraphQL type system on top of Ruby's type system, which results in a pretty complex inheritance chain. To summarize:
* `ObjectType.new` returns a new class: a subclass of `ObjectClass` which is tied to a graphql type. This is an "ObjectType". This class is never in practice instantiated since it has no fields (other than in tests). These are usually assigned to a constant and so have a `name`.
* `ObjectType#define_class` makes another subclass from that object type and ties it to how this type is used in a query: the query definition, the fields being included, their types (this ends up recursive), etc. These have methods defined for every field they include. A new class is created here each time a type is referenced in a query. These are usually anonymous classes.
* `new` is called on those "defined classes", to build these `ObjectClass` instances.

This PR changes how this works.

* The "ObjectType"s, returned by `ObjectType.new`, can now be instantiated, but expect to be passed information about the definiton: fields defined on it, types, spreads, etc.
* `ObjectType#define_class` now returns a new `WithDefinition` instance. This is a plain Ruby object, instead of a subclass. It implements `#new` which will make an ObjectClass with the correct information from this definition.

This is made possible by removing the method definitions for fields, and instead relying on `method_missing` every time a field is accessed.