@jhawthorn 2020-08-04
> The tests all pass and the code looks good to me. The only question I have about the design is about method_missing: it _used to_ be that `method_missing` wasn't recommended for hot paths because it was a lot slower than normal method calls (because it couldn't use the method cache, maybe?). Is that still a consideration here? Or, has that improved in Ruby? Or do you think it'll be worth it for the memory improvements?

Good call! @dinahshi and I took a look at this.

We started by benchmarking an attribute on one of these object types. It's a bit slower, but both are fast enough that I think the trade off is worth it: over 1M i/s. I think it's probably somewhat rare that the results of the client query are called in a loop like this, and more often the querying and casting costs will outweigh this access.

<details>
<summary>Benchmark script</summary>

```
require "graphql/client"

class PersonType < GraphQL::Schema::Object
  field :name, String, null: true
end

class QueryType < GraphQL::Schema::Object
  field :me, PersonType, null: false
  def me
    {}
  end
end

class Schema < GraphQL::Schema
  query(QueryType)
end
client = GraphQL::Client.new(schema: Schema, execute: Schema)

PersonQuery = client.parse(<<~GRAPHQL)
{
  me {
    name
  }
}
GRAPHQL

result = PersonQuery.new({"me" => {"name" => "John"}})
me = result.me

require "benchmark/ips"
Benchmark.ips do |x|
  x.report "attribute read" do
    me.name
  end
end
```

</details>

**Before**
```
$ be ruby -Ilib benchmark_attr.rb
Warming up --------------------------------------
      attribute read   216.900k i/100ms
Calculating -------------------------------------
      attribute read      2.284M (±13.6%) i/s -     11.279M in   5.086265s
```


**After**
```
$ be ruby -Ilib benchmark_attr.rb
Warming up --------------------------------------
      attribute read   139.978k i/100ms
Calculating -------------------------------------
      attribute read      1.645M (± 9.0%) i/s -      8.259M in   5.068651s
```


w/r/t `method_missing` it is definitely a little bit slower, but still relatively fast. Our method calls were also already made slower because they're dynamic `define_method`s being passed a Proc and needing a closure.

<details>
<summary> Benchmark script </summary>

```
# frozen_string_literal: true

class PORO
  def name
    123
  end
end

class Dynamic
  p = -> { 123 }
  define_method(:name, &p)
end

class MM
  def method_missing(name, *args)
    123
  end
end

poro = PORO.new
mm = MM.new
dynamic = Dynamic.new

require "benchmark/ips"
Benchmark.ips do |x|
  x.report "plain method" do
    poro.name
  end

  x.report "method_missing" do
    mm.name
  end

  x.report "define_method" do
    dynamic.name
  end
end
```

</details>


```
$ be ruby benchmark_method_missing.rb
Warming up --------------------------------------
        plain method     1.165M i/100ms
      method_missing   483.899k i/100ms
       define_method   795.723k i/100ms
Calculating -------------------------------------
        plain method     13.733M (± 9.1%) i/s -     68.719M in   5.052126s
      method_missing      5.034M (±10.0%) i/s -     25.163M in   5.056971s
       define_method      7.629M (±12.4%) i/s -     37.399M in   5.004813s
```